+++ django_enum/choices.py
@@ -9,7 +9,11 @@ from typing import Any, List, Optional, Tuple, Type
 from django.db.models import Choices
 from django.db.models import IntegerChoices as DjangoIntegerChoices
 from django.db.models import TextChoices as DjangoTextChoices
-from django.db.models.enums import ChoicesMeta
+
+try:
+    from django.db.models.enums import ChoicesType
+except ImportError:  # pragma: no cover
+    from django.db.models.enums import ChoicesMeta as ChoicesType
 
 
 def choices(enum: Optional[Type[Enum]]) -> List[Tuple[Any, str]]:
@@ -87,7 +91,7 @@ try:
     from enum_properties import EnumPropertiesMeta, SymmetricMixin
 
 
-    class DjangoEnumPropertiesMeta(EnumPropertiesMeta, ChoicesMeta):
+    class DjangoEnumPropertiesMeta(EnumPropertiesMeta, ChoicesType):
         """
         A composite meta class that combines Django's Choices metaclass with
         enum-properties metaclass. This metaclass will add Django's expected
@@ -104,7 +108,7 @@ try:
         _symmetric_builtins_ = ['name', 'label']
 
 
-    class TextChoices(
+    class TextChoices(  # pylint: disable=too-many-ancestors
         DjangoSymmetricMixin,
         DjangoTextChoices,
         metaclass=DjangoEnumPropertiesMeta
@@ -118,7 +122,7 @@ try:
             return DjangoTextChoices.__hash__(self)
 
 
-    class IntegerChoices(
+    class IntegerChoices(  # pylint: disable=too-many-ancestors
         DjangoSymmetricMixin,
         DjangoIntegerChoices,
         metaclass=DjangoEnumPropertiesMeta
@@ -146,6 +150,9 @@ try:
         def __hash__(self):
             return float.__hash__(self)
 
+        def __str__(self):
+            return str(self.value)
+
 
 except (ImportError, ModuleNotFoundError):
 
@@ -162,7 +169,7 @@ except (ImportError, ModuleNotFoundError):
     DjangoSymmetricMixin = MissingEnumProperties  # type: ignore
 
 
-    class DjangoEnumPropertiesMeta(ChoicesMeta):  # type: ignore
+    class DjangoEnumPropertiesMeta(ChoicesType):  # type: ignore
         """
         Throw error if metaclass is used without enum-properties