+++ django_enum/fields.py
@@ -27,6 +27,13 @@ from django.db.models import (
     SmallIntegerField,
 )
 from django.db.models.query_utils import DeferredAttribute
+
+try:
+    from django.db.models.expressions import DatabaseDefault
+except ImportError:  # pragma: no cover
+    class DatabaseDefault:  # type: ignore
+        """Spoof DatabaseDefault for Django < 5.0"""
+
 from django_enum.choices import choices, values
 from django_enum.forms import EnumChoiceField, NonStrictSelect
 
@@ -56,7 +63,11 @@ class ToPythonDeferredAttribute(DeferredAttribute):
 
     def __set__(self, instance: Model, value: Any):
         try:
-            instance.__dict__[self.field.name] = self.field.to_python(value)
+            instance.__dict__[self.field.name] = (
+                value
+                if isinstance(value, DatabaseDefault) else
+                self.field.to_python(value)
+            )
         except (ValidationError, ValueError):
             # Django core fields allow assignment of any value, we do the same
             instance.__dict__[self.field.name] = value
@@ -83,18 +94,9 @@ class EnumMixin(
     enum: Optional[Type[Enum]] = None
     strict: bool = True
     coerce: bool = True
+    primitive: Type[Any]
 
     descriptor_class = ToPythonDeferredAttribute
-
-    def _coerce_to_value_type(self, value: Any) -> Enum:
-        """Coerce the value to the enumerations value type"""
-        # note if enum type is int and a floating point is passed we could get
-        # situations like X.xxx == X - this is acceptable
-        if self.enum:
-            return type(values(self.enum)[0])(value)
-        # can't ever reach this - just here to make type checker happy
-        return value  # pragma: no cover
-
 
     def __init__(
             self,
@@ -121,16 +123,15 @@ class EnumMixin(
         and non-strict, coercion to enum's primitive type will be done,
         otherwise a ValueError is raised.
         """
-        if (
-            (self.coerce or force)
-            and self.enum is not None
-            and not isinstance(value, self.enum)
-        ):
+        if self.enum is None:
+            return value
+
+        if (self.coerce or force) and not isinstance(value, self.enum):
             try:
                 value = self.enum(value)
             except (TypeError, ValueError):
                 try:
-                    value = self._coerce_to_value_type(value)
+                    value = self.primitive(value)
                     value = self.enum(value)
                 except (TypeError, ValueError):
                     try:
@@ -138,13 +139,26 @@ class EnumMixin(
                     except KeyError as err:
                         if self.strict or not isinstance(
                             value,
-                            type(values(self.enum)[0])
+                            self.primitive
                         ):
                             raise ValueError(
                                 f"'{value}' is not a valid "
                                 f"{self.enum.__name__} "
                                 f"required by field {self.name}."
                             ) from err
+        elif (
+            not self.coerce and
+            not isinstance(value, self.primitive) and
+            not isinstance(value, self.enum)
+        ):
+            try:
+                return self.primitive(value)
+            except (TypeError, ValueError) as err:
+                raise ValueError(
+                    f"'{value}' is not coercible to {self.primitive.__name__} "
+                    f"required by field {self.name}."
+                ) from err
+
         return value
 
     def deconstruct(self) -> Tuple[str, str, List, dict]:
@@ -306,6 +320,8 @@ class EnumCharField(EnumMixin, CharField):
     A database field supporting enumerations with character values.
     """
 
+    primitive = str
+
     def __init__(self, *args, enum=None, **kwargs):
         kwargs.setdefault(
             'max_length',
@@ -320,12 +336,16 @@ class EnumCharField(EnumMixin, CharField):
 class EnumFloatField(EnumMixin, FloatField):
     """A database field supporting enumerations with floating point values"""
 
+    primitive = float
+
 
 class EnumSmallIntegerField(EnumMixin, SmallIntegerField):
     """
     A database field supporting enumerations with integer values that fit into
     2 bytes or fewer
     """
+
+    primitive = int
 
 
 class EnumPositiveSmallIntegerField(EnumMixin, PositiveSmallIntegerField):
@@ -334,12 +354,15 @@ class EnumPositiveSmallIntegerField(EnumMixin, PositiveSmallIntegerField):
     values that fit into 2 bytes or fewer
     """
 
+    primitive = int
 
 class EnumIntegerField(EnumMixin, IntegerField):
     """
     A database field supporting enumerations with integer values that fit into
     32 bytes or fewer
     """
+
+    primitive = int
 
 
 class EnumPositiveIntegerField(EnumMixin, PositiveIntegerField):
@@ -348,6 +371,8 @@ class EnumPositiveIntegerField(EnumMixin, PositiveIntegerField):
     values that fit into 32 bytes or fewer
     """
 
+    primitive = int
+
 
 class EnumBigIntegerField(EnumMixin, BigIntegerField):
     """
@@ -355,12 +380,16 @@ class EnumBigIntegerField(EnumMixin, BigIntegerField):
     64 bytes or fewer
     """
 
+    primitive = int
+
 
 class EnumPositiveBigIntegerField(EnumMixin, PositiveBigIntegerField):
     """
     A database field supporting enumerations with positive (but signed) integer
     values that fit into 64 bytes or fewer
     """
+
+    primitive = int
 
 
 class _EnumFieldMetaClass(type):